requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.4
urllib3==2.5.0
retry==0.9.2

google-cloud-storage==2.15.0
